feat: #2 ajouter memoire de contexte à la question ---
tested :
curl --location --request GET 'localhost:8080/scAlbert/ask' \
--header 'Content-Type: application/json' \
--data '{
    "goal" : "l'\''objectif : creer un chatbot dans mon application",
    "question": "question : comment faire un chatbot basé sur l'\''ia",
    "contexte": "contexte :the project you are working on is a web app with VueJS in the front et backend : (java spring, langchain4j)",

    "role" : "You are a senior developer explaining to another senior developer, ",
    "precision":""
}'
