+++ src/main/java/io/tintin/ia/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package io.tintin.ia;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
