--- src/main/java/org/example/ChatLanguageModelController.java
@@ -1,21 +1,0 @@
-package org.example;
-
-import dev.langchain4j.model.chat.ChatLanguageModel;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-class ChatLanguageModelController {
-
-    ChatLanguageModel chatLanguageModel;
-
-    ChatLanguageModelController(ChatLanguageModel chatLanguageModel) {
-        this.chatLanguageModel = chatLanguageModel;
-    }
-
-    @GetMapping("/model")
-    public String model(@RequestParam(value = "message", defaultValue = "Hello") String message) {
-        return chatLanguageModel.generate(message);
-    }
-}
